fix(CI): request libwebkit2gtk=2.44.0-2 to fix AppImages on Fedora ~43
The CI builds the GUI on ubuntu noble, which contains
  2.44.0-2
in universe and
  2.48.7-0ubuntu0.24.04.2
in security. Naturally, the bigger version is picked by default.

The same thing happens when building on bookworm (2.48.1-2~deb12u1).

This version causes the GUI to crash just after creating a blank window,
yielding no logs and stderr of
  Could not create default EGL display: EGL_BAD_PARAMETER. Aborting...

This is a known issue in the Tauri/AppImage world,
and downgrading to 2.44 is likewise a known fix, cf.
  https://github.com/gitbutlerapp/gitbutler/issues/5282
and all the referencing issues.

On bookworm this can be fixed by installing 2.44.3-1
off snapshot.d.o at 20240909T145608Z.

On noble, we can just ask for the older version, since both are visible.

Only pin on noble: we only ever request noble,
but some third-party workflows are on jammy.
The version doesn't matter there.

Closes #594
Closes #582